requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
astunparse==1.6.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.2
executing==2.0.1
facenet-pytorch==2.6.0
filelock==3.9.0
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2023.4.0
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.62.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.7
httpx==0.27.2
idna==3.7
importlib_metadata==8.4.0
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.2
jiter==0.7.1
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
nest_asyncio==1.6.0
networkx==3.2.1
numpy==1.26.3
oauthlib==3.2.2
openai==1.54.4
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.0
pandas==2.2.2
parso==0.8.4
pickleshare==0.7.5
pillow==10.2.0
pip==23.3.1
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==3.19.6
psutil==5.9.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.1.4
pyreadline3==3.5.4
python-dateutil==2.9.0
python-dotenv==1.0.1
pytz==2024.1
pywin32==305.1
pyzmq==25.1.2
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.20.0
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
setuptools==68.2.2
six==1.16.0
sniffio==1.3.1
stack-data==0.6.2
sympy==1.12
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.9.20
torch==2.2.2+cu121
torchaudio==2.2.2+cu121
torchinfo==1.8.0
torchsummary==1.5.1
torchvision==0.17.2+cu121
tornado==6.2
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.2
wheel==0.41.2
wrapt==1.16.0
zipp==3.20.0
